+++ src/ui/components/charts/index.js
@@ -1,4 +1,6 @@
 module.exports = {
 	...require('./bar-chart.js'),
 	...require('./line-chart.js'),
+	...require('./pie-chart.js'),
+	...require('./donut-chart.js'),
 }