+++ src/ui/style/style-names.js
@@ -22,10 +22,45 @@ const styleNames = (values) => {
 }
 
 const initDefaultStyleNames = () => {
+	const text = {
+		fontWeight: 'var:fontWeight',
+		fg: 'var:fg',
+		fontSize: 'var:fontSize',
+		fontFamily: 'var:fontFamily',
+	}
+
 	const names = {
 		row: { direction: 'row' },
 		column: { direction: 'column' },
 		hug: { size: 'hug' },
+		radius: { radius: 'var:radius' },
+		card: {
+			radius: 'var:radius',
+			bg: 'var:componentBg',
+			padding: 'var:u2',
+		},
+		body: { bg: 'var:bg' },
+		text,
+		h0: {
+			...text,
+			fontSize: 'var:h0FontSize',
+		},
+		h1: {
+			...text,
+			fontSize: 'var:h1FontSize',
+		},
+		h2: {
+			...text,
+			fontSize: 'var:h2FontSize',
+		},
+		h3: {
+			...text,
+			fontSize: 'var:h3FontSize',
+		},
+		link: {
+			...text,
+			fg: 'var:accent',
+		},
 	}
 
 	const items = [1, 2, 3, 4, 5]
@@ -35,6 +70,24 @@ const initDefaultStyleNames = () => {
 		names[`p_${item}`] = { padding: `var:u${item}`}
 		names[`size_${item}`] = { size: item }
 	}))
+
+	const textAlignments = ['start', 'end', 'center']
+	const textBaselines = [
+		'top',
+		'hanging',
+		'middle',
+		'alphabetic',
+		'ideographic',
+		'bottom',
+	]
+
+	textAlignments.forEach((ta) => {
+		names[`ta_${ta}`] = { textAlign: ta }
+	})
+
+	textBaselines.forEach((tb) => {
+		names[`tb_${tb}`] = { textBaseline: tb }
+	})
 
 	styleNames(names)
 }